Added tests for few models

+++ db/migrate/20200921052228_create_anything_god_records.rb
@@ -1,7 +1,7 @@
 class CreateAnythingGodRecords < ActiveRecord::Migration[5.2]
   def change
     create_table :at_god_records do |t|
-      t.integer :collection_id, null: false
+      t.bigint :collection_id, null: false
       t.jsonb :anything_data, default: '{}'
 
       t.timestamps null: false
